+++ hypothesis-python/tests/nocover/test_simple_strings.py
@@ -17,5 +17,13 @@
 
 from __future__ import division, print_function, absolute_import
 
-__version_info__ = (3, 70, 0)
-__version__ = '.'.join(map(str, __version_info__))
+import unicodedata
+
+from hypothesis import given, settings
+from hypothesis.strategies import text
+
+
+@given(text(min_size=1, max_size=1))
+@settings(max_examples=2000)
+def test_does_not_generate_surrogates(t):
+    assert unicodedata.category(t) != u'Cs'