+++ hypothesis-ruby/lib/hypothesis.rb
@@ -122,7 +122,7 @@ module Hypothesis
   # of the block is a *test case*.
   # A test case has three important features:
   #
-  # * *givens* are the result of a call to self.given, and are the
+  # * *givens* are the result of a call to self.any, and are the
   #   values that make up the test case. These might be values such
   #   as strings, integers, etc. or they might be values specific to
   #   your application such as a User object.